Fix memory issues in SmoothingBuffer. Fixed warnings about non-const strings.

+++ include/SmoothingBuffer.h
@@ -5,7 +5,7 @@
 
 class SmoothingBuffer {
     public:
-        SmoothingBuffer(int size);
+        SmoothingBuffer(size_t size);
         void storeValue(float value);
         float getAverage();
