+++ src/SmoothingBuffer.cpp
@@ -1,7 +1,7 @@
 #include <Arduino.h>
 #include "SmoothingBuffer.h"
 
-SmoothingBuffer::SmoothingBuffer(int size) {
+SmoothingBuffer::SmoothingBuffer(size_t size) {
     bufferSize = size;
     values = new float[bufferSize];
     for(int i = 0; i < bufferSize; i++) {
@@ -24,7 +24,6 @@ float SmoothingBuffer::getAverage() {
     float total = 0.0;
     for(int i = 0; i < valuesInBuffer; i++) {
         total += values[i];
-        // yield();
     }
     return total / (float) valuesInBuffer;
 }
